Commited intermediate work because initial work on Euclidean distances counting has strange behavior. Initially I tried to use floats but I catch a problem when every two distances was duplicate. This is a wrong result so after investigating I found that disabling -O3 option will fix this option. Unfortunatelly I need optimized execution so I switched from float to double base type. Double works correct with -O3 compiler option.

+++ src/histogramMain.c
@@ -23,7 +23,7 @@ int main(int argv, char* argc[]){
 	const int nTrials = 10;
 	//const int nTrials = 4;
 	
-	float* age = (float*) _mm_malloc(sizeof(int)*n, 64);
+	float* age = (float*) _mm_malloc(sizeof(float)*n, 64);
 	
 	int refGroup[m];
